1 - Research Tool/requirements.txt
streamlit~=1.34.0
langchain~=0.2.0
python-dotenv~=1.0.1